Changed structure/names of packages
