+++ build.gradle
@@ -18,7 +18,7 @@ apply plugin: 'spring-boot'
 apply plugin: 'io.spring.dependency-management'
 
 jar {
-    baseName = 'gog'
+    baseName = 'yagga'
     version = '0.0.1-SNAPSHOT'
 }
 sourceCompatibility = 1.8
