--- gog-config.properties
@@ -1,2 +1,0 @@
-repositories.paths=/home/user/all_repositories/directory_with_some_git_repositories,\
-  /home/user/all_repositories/other_directory_with_some_git_repositories
