--- src/main/java/com/github/yu55/gog/Application.java
@@ -1,12 +1,0 @@
-package com.github.yu55.gog;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
