--- src/main/java/com/github/yu55/gog/core/model/AnnotateRequest.java
@@ -1,16 +1,0 @@
-package com.github.yu55.gog.core.model;
-
-public class AnnotateRequest {
-
-    private String repository;
-
-    private String file;
-
-    public String getRepository() {
-        return repository;
-    }
-
-    public String getFile() {
-        return file;
-    }
-}
