--- src/main/java/com/github/yu55/gog/handler/git/command/GitCommand.java
@@ -1,10 +1,0 @@
-package com.github.yu55.gog.handler.git.command;
-
-import org.apache.commons.exec.CommandLine;
-
-public interface GitCommand {
-
-    String COMMAND = "git";
-
-    CommandLine getCommandLine();
-}
