--- src/main/java/com/github/yu55/gog/handler/git/command/GitCommandOutputLine.java
@@ -1,14 +1,0 @@
-package com.github.yu55.gog.handler.git.command;
-
-public class GitCommandOutputLine {
-
-    private String line;
-
-    public GitCommandOutputLine(String line) {
-        this.line = line;
-    }
-
-    public String getLine() {
-        return line;
-    }
-}
