+++ src/main/java/org/yu55/yagga/core/annotate/model/AnnotateRequest.java
@@ -1,0 +1,16 @@
+package org.yu55.yagga.core.annotate.model;
+
+public class AnnotateRequest {
+
+    private String repository;
+
+    private String file;
+
+    public String getRepository() {
+        return repository;
+    }
+
+    public String getFile() {
+        return file;
+    }
+}
