+++ src/main/java/org/yu55/yagga/handler/git/command/common/GitCommand.java
@@ -1,0 +1,10 @@
+package org.yu55.yagga.handler.git.command.common;
+
+import org.apache.commons.exec.CommandLine;
+
+public interface GitCommand {
+
+    String COMMAND = "git";
+
+    CommandLine getCommandLine();
+}
