+++ src/main/java/org/yu55/yagga/handler/git/command/common/GitCommandOutputLine.java
@@ -1,0 +1,14 @@
+package org.yu55.yagga.handler.git.command.common;
+
+public class GitCommandOutputLine {
+
+    private String line;
+
+    public GitCommandOutputLine(String line) {
+        this.line = line;
+    }
+
+    public String getLine() {
+        return line;
+    }
+}
