+++ src/main/resources/static/index.html
@@ -2,7 +2,7 @@
 <html lang="en" ng-app="gog">
 <head>
     <meta charset="UTF-8">
-    <title>The GOG project</title>
+    <title>yagga</title>
     <script src="//cdnjs.cloudflare.com/ajax/libs/angular.js/1.2.28/angular.min.js"></script>
     <script>
         var nameApp = angular.module('gog', []);
@@ -68,7 +68,7 @@
     </script>
 </head>
 <body ng-controller="gogCtrl">
-<h1 style="font-family: monospace">$ Git Online Grep</h1>
+<h1 style="font-family: monospace">$ YAGGA</h1>
 
 Provide text to search for:
 <br/>
