+++ yagga-config.properties
@@ -1,0 +1,3 @@
+repositories.paths=/home/user/all_repositories/directory_with_some_git_repositories,\
+  /home/user/all_repositories/other_directory_with_some_git_repositories
+
